+++ src/modules.c
@@ -632,8 +632,8 @@ drop_all_hooks(aModule *owner)
         if(hk->owner == owner)
         {
             sendto_realops_lev(DEBUG_LEV, "Module cleanup: removing hook [%s]"
-                            " for opaque %d", get_texthooktype(hk->hooktype), 
-                            (int) owner);
+                            " for opaque %lu", get_texthooktype(hk->hooktype), 
+                            (u_long) owner);
 
             hooklist = get_hooklist((enum c_hooktype) hk->hooktype);
 
@@ -653,8 +653,8 @@ bircmodule_add_hook(enum c_hooktype hooktype, void *opaque, void *funcptr)
 
     if(!(owner = find_module_opaque(opaque)))
     {
-        sendto_realops_lev(DEBUG_LEV, "Module tried to add hooktype %d with"
-                         " unknown opaque 0x%x", (int) hooktype, (int) opaque);
+        sendto_realops_lev(DEBUG_LEV, "Module tried to add hooktype %lu with"
+                         " unknown opaque 0x%x", (u_long) hooktype, (u_long) opaque);
         return NULL;
     }
